objects: move more TR1 objects to TRX
