+++ data/tr2/ship/cfg/tr2-gm/gameflow.json5
@@ -199,5 +199,6 @@
         "exposurebar_color",
         "exposurebar_location",
         "exposurebar_show_mode",
+        "change_pierre_spawn",
     ],
 }
