+++ data/tr2/ship/cfg/tr2/gameflow.json5
@@ -644,5 +644,6 @@
         "exposurebar_color",
         "exposurebar_location",
         "exposurebar_show_mode",
+        "change_pierre_spawn",
     ],
 }
