+++ src/libtrx/config/map.def
@@ -17,6 +17,7 @@ X_CFG_BOOL(debug.enable_debug_spheres,               false)
 X_CFG_BOOL(debug.enable_debug_triggers,              false)
 X_CFG_BOOL(debug.enable_invulnerability,             false)
 X_CFG_BOOL(debug.enable_review_markers,              false)
+X_CFG_BOOL(gameplay.change_pierre_spawn,             true)
 X_CFG_BOOL(gameplay.disable_healing_between_levels,  false)
 X_CFG_BOOL(gameplay.disable_trex_collision,          false)
 X_CFG_BOOL(gameplay.enable_auto_item_selection,      true)
