+++ src/libtrx/config/map_tr1.def
@@ -3,7 +3,6 @@ X_CFG_BOOL(audio.fix_secrets_killing_music,    true)
 X_CFG_BOOL(audio.fix_speeches_killing_music,   true)
 X_CFG_BOOL(audio.fix_tihocan_secret_sound,     true)
 X_CFG_BOOL(audio.load_music_triggers,          true)
-X_CFG_BOOL(gameplay.change_pierre_spawn,       true)
 X_CFG_BOOL(gameplay.disable_magnums,           false)
 X_CFG_BOOL(gameplay.disable_medpacks,          false)
 X_CFG_BOOL(gameplay.disable_shotgun,           false)
