+++ src/libtrx/game/objects/creatures/mutant.c
@@ -1,13 +1,13 @@
 #include "game/objects/creatures/mutant.h"
 
+#include "game/carrier.h"
 #include "game/creature.h"
 #include "game/lara.h"
+#include "game/pathing.h"
+#include "game/random.h"
+#include "game/sound.h"
 #include "game/spawn.h"
-
-#include <libtrx/game/carrier.h>
-#include <libtrx/game/random.h>
-#include <libtrx/game/sound.h>
-#include <libtrx/utils.h>
+#include "utils.h"
 
 #define FLYER_CHARGE_DAMAGE 100
 #define FLYER_LUNGE_DAMAGE 150
