+++ src/libtrx/game/spawn.c
@@ -3,9 +3,29 @@
 #include "game/collision.h"
 #include "game/effects.h"
 #include "game/lara.h"
+#include "game/math.h"
 #include "game/random.h"
 #include "game/rooms.h"
 #include "game/sound.h"
+
+#define M_SHARD_SPEED 250
+#define M_ROCKET_SPEED 220
+
+static void M_ShootAtLara(EFFECT *const effect)
+{
+    const ITEM *const lara_item = Lara_GetItem();
+    const BOUNDS_16 *const bounds = Item_GetBoundsAccurate(lara_item);
+    const int32_t x = lara_item->pos.x - effect->pos.x;
+    const int32_t z = lara_item->pos.z - effect->pos.z;
+    const int32_t y = lara_item->pos.y - effect->pos.y
+        + (bounds->max.y + (bounds->min.y - bounds->max.y) * 3 / 4);
+
+    const int32_t dist = Math_Sqrt(SQUARE(x) + SQUARE(z));
+    effect->rot.x = -Math_Atan(dist, y);
+    effect->rot.y = Math_Atan(z, x);
+    effect->rot.x += (Random_GetControl() - 0x4000) / 0x40;
+    effect->rot.y += (Random_GetControl() - 0x4000) / 0x40;
+}
 
 void Spawn_Splash(const ITEM *const item)
 {
@@ -145,4 +165,50 @@ int16_t Spawn_GunMiss(
     };
     Spawn_Ricochet(&pos);
     return Spawn_GunShot(x, y, z, speed, y_rot, room_num);
+}
+
+int16_t Spawn_ShardGun(
+    int32_t x, int32_t y, int32_t z, int16_t speed, int16_t y_rot,
+    int16_t room_num)
+{
+    int16_t effect_num = Effect_Create(room_num);
+    if (effect_num != NO_EFFECT) {
+        EFFECT *effect = Effect_Get(effect_num);
+        effect->room_num = room_num;
+        effect->pos.x = x;
+        effect->pos.y = y;
+        effect->pos.z = z;
+        effect->rot.x = 0;
+        effect->rot.y = y_rot;
+        effect->rot.z = 0;
+        effect->object_id = O_MISSILE_2;
+        effect->frame_num = 0;
+        effect->speed = M_SHARD_SPEED;
+        effect->shade = 3584;
+        M_ShootAtLara(effect);
+    }
+    return effect_num;
+}
+
+int16_t Spawn_RocketGun(
+    int32_t x, int32_t y, int32_t z, int16_t speed, int16_t y_rot,
+    int16_t room_num)
+{
+    int16_t effect_num = Effect_Create(room_num);
+    if (effect_num != NO_EFFECT) {
+        EFFECT *effect = Effect_Get(effect_num);
+        effect->room_num = room_num;
+        effect->pos.x = x;
+        effect->pos.y = y;
+        effect->pos.z = z;
+        effect->rot.x = 0;
+        effect->rot.y = y_rot;
+        effect->rot.z = 0;
+        effect->object_id = O_MISSILE_3;
+        effect->frame_num = 0;
+        effect->speed = M_ROCKET_SPEED;
+        effect->shade = SHADE_NEUTRAL;
+        M_ShootAtLara(effect);
+    }
+    return effect_num;
 }
