+++ src/libtrx/game/ui/dialogs/setting_tabs/gameplay_general.def
@@ -34,9 +34,9 @@ X_UI_CFG_BOOL_INV(gameplay.revert_to_pistols,      SETTING_REMEMBER_GUNS_BETWEEN
 
 #if TR_VERSION == 1
 X_UI_CFG_BOOL(gameplay.restore_ps1_enemies,        SETTING_RESTORE_PS1_ENEMIES)
-X_UI_CFG_BOOL(gameplay.change_pierre_spawn,        SETTING_CHANGE_PIERRE_SPAWN)
 #endif
 
+X_UI_CFG_BOOL(gameplay.change_pierre_spawn,        SETTING_CHANGE_PIERRE_SPAWN)
 X_UI_CFG_BOOL(gameplay.disable_trex_collision,     SETTING_DISABLE_TREX_COLLISION)
 X_UI_CFG_BOOL(gameplay.enable_enemy_rotation,      SETTING_ENABLE_ENEMY_ROTATION)
