+++ src/libtrx/include/libtrx/game/objects/creatures/mutant.h
