+++ src/libtrx/include/libtrx/game/spawn.h
@@ -23,3 +23,10 @@ int16_t Spawn_GunHit(
 int16_t Spawn_GunMiss(
     int32_t x, int32_t y, int32_t z, int16_t speed, int16_t y_rot,
     int16_t room_num);
+
+int16_t Spawn_ShardGun(
+    int32_t x, int32_t y, int32_t z, int16_t speed, int16_t y_rot,
+    int16_t room_num);
+int16_t Spawn_RocketGun(
+    int32_t x, int32_t y, int32_t z, int16_t speed, int16_t y_rot,
+    int16_t room_num);
