+++ src/libtrx/meson.build
@@ -238,14 +238,24 @@ sources = [
   'game/objects/common.c',
   'game/objects/creatures/ape.c',
   'game/objects/creatures/bacon_lara.c',
+  'game/objects/creatures/baldy.c',
   'game/objects/creatures/bat.c',
   'game/objects/creatures/bear.c',
+  'game/objects/creatures/centaur.c',
+  'game/objects/creatures/centaur_statue.c',
+  'game/objects/creatures/cowboy.c',
   'game/objects/creatures/crocodile.c',
+  'game/objects/creatures/larson.c',
   'game/objects/creatures/lion.c',
   'game/objects/creatures/mummy.c',
+  'game/objects/creatures/mutant.c',
+  'game/objects/creatures/natla.c',
+  'game/objects/creatures/pierre.c',
   'game/objects/creatures/pod.c',
   'game/objects/creatures/raptor.c',
   'game/objects/creatures/rat.c',
+  'game/objects/creatures/skate_kid.c',
+  'game/objects/creatures/torso.c',
   'game/objects/creatures/trex.c',
   'game/objects/creatures/wolf.c',
   'game/objects/draw.c',
@@ -257,6 +267,7 @@ sources = [
   'game/objects/effects/explosion.c',
   'game/objects/effects/flame.c',
   'game/objects/effects/gun_flash.c',
+  'game/objects/effects/missile.c',
   'game/objects/effects/natla_gun.c',
   'game/objects/effects/pickup_aid.c',
   'game/objects/effects/ricochet.c',
