+++ src/tr1/game/creature.h
@@ -1,7 +1,3 @@
 #pragma once
 
 #include <libtrx/game/creature.h>
-
-bool Creature_ShootAtLara(
-    ITEM *item, int32_t distance, BITE *gun, int16_t extra_rotation,
-    int16_t damage);
