+++ src/tr1/game/item_actions/bubbles.c
@@ -1,11 +1,11 @@
 #include "game/item_actions/bubbles.h"
 
-#include "game/spawn.h"
-
 #include <libtrx/game/collision.h>
 #include <libtrx/game/lara.h>
 #include <libtrx/game/random.h>
+#include <libtrx/game/rooms.h>
 #include <libtrx/game/sound.h>
+#include <libtrx/game/spawn.h>
 
 void ItemAction_Bubbles(ITEM *const item)
 {
