--- src/tr1/game/spawn.h
@@ -1,13 +1,0 @@
-#pragma once
-
-#include "global/types.h"
-
-#include <libtrx/game/spawn.h>
-
-int16_t Spawn_ShardGun(
-    int32_t x, int32_t y, int32_t z, int16_t speed, int16_t y_rot,
-    int16_t room_num);
-
-int16_t Spawn_RocketGun(
-    int32_t x, int32_t y, int32_t z, int16_t speed, int16_t y_rot,
-    int16_t room_num);
