+++ src/tr1/meson.build
@@ -134,18 +134,7 @@ sources = [
   'game/level.c',
   'game/los.c',
   'game/objects/common.c',
-  'game/objects/creatures/baldy.c',
-  'game/objects/creatures/centaur.c',
-  'game/objects/creatures/centaur_statue.c',
-  'game/objects/creatures/cowboy.c',
   'game/objects/creatures/cutscene_player.c',
-  'game/objects/creatures/larson.c',
-  'game/objects/creatures/mutant.c',
-  'game/objects/creatures/natla.c',
-  'game/objects/creatures/pierre.c',
-  'game/objects/creatures/skate_kid.c',
-  'game/objects/creatures/torso.c',
-  'game/objects/effects/missile.c',
   'game/objects/general/earthquake.c',
   'game/objects/general/pickup.c',
   'game/objects/general/save_crystal.c',
@@ -162,7 +151,6 @@ sources = [
   'game/savegame/savegame_bson.c',
   'game/savegame/savegame_legacy.c',
   'game/shell/common.c',
-  'game/spawn.c',
   'game/stats/common.c',
   'game/ui/dialogs/stats.c',
   'global/enum_map.c',
